moved sign in to controller
