+++ src/app/routers/authRouter.ts
@@ -10,6 +10,9 @@ import Validate from '../drivers/Validate';
 // middleware
 import {authenticate} from '../middleware/authenticate';
 
+// controllers
+import userController from '../controllers/userController';
+
 // models
 import User from '../models/User';
 
@@ -84,66 +87,7 @@ router.post('/create-user', async function (req: Request, res: Response) {
 
 // sign in and get jwt token
 router.post('/', async function (req: Request, res: Response) {
-	// ensure email and password were passed
-	const missingInputErrors = Validate.require(['email', 'password'], req.body);
-
-	// return any errors
-	if (missingInputErrors.length) {
-		res
-			.status(400)
-			.send(<http.Response> {
-				success: false,
-				errors: missingInputErrors
-			});
-
-		return;
-	}
-
-	const {email, password} = req.body;
-
-	// grab user from db
-	const user = await User.findOne({email});
-
-	// return error if user not found
-	if (!user) {
-		res
-			.status(400)
-			.send(<http.Response> {
-				success: false,
-				errors: [<http.ResponseError> {
-					title: 'Invalid E-mail',
-					detail: 'E-mail was not tied to any existing user.',
-					httpStatus: 400
-				}]
-			});
-
-		return;
-	}
-
-	// return error if password does not match
-	if (!Auth.stringMatchesHash(password, user.get('password'))) {
-		res
-			.status(400)
-			.send(<http.Response> {
-				success: false,
-				errors: [<http.ResponseError> {
-					title: 'Invalid Password',
-					httpStatus: 400
-				}]
-			});
-
-		return;
-	}
-
-	const token = Auth.generateToken({
-		id: user._id,
-		email
-	});
-
-	res.send(<http.Response> {
-		success: true,
-		data: {token}
-	});
+	return userController.signIn(req, res);
 });
 
 // test authed endpoint
